### 2025-05-09 #8187 @editorialbot
[PRE REVIEW]: Mantaray: A Rust Package for Ray Tracing Ocean Surface Gravity Waves
**Submitting author:** @biavillasboas (<a href="http://orcid.org/0000-0001-6767-6556">Ana Beatriz Villas Bôas</a>)
**Repository:** https://github.com/mines-oceanography/mantaray
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
