### 2025-05-11 #8187 @biavillasboas
Thank you, @kthyng!

It seems like the only issue so far is with DOIs in the references. I'll take a look at that. In the meantime, some folks who we think would be great reviewers are:

- Gaute Hope (gauteh): Oceanographer with ample knowledge of Rust. Has co-authored a package similar to ours, but in pure Python.
- Milan Curcic (milancurcic): Oceanographer with strong background in ocean waves and lots of experience/contributions in the open source community.
- Deepak Cherian (dcherian): Oceanographer with strong background in open source software development and Rust.
