+++ Cargo.toml
@@ -31,7 +31,7 @@ ndarray = { version = "0.16.1", features = ["rayon"] }
 netcdf3 = "0.5.2"
 netcdf = {version = "0.11.0", features = ["static", "ndarray"]}
 ode_solvers = "0.4.0"
-pyo3 = { version = "0.24.2" }
+pyo3 = { version = "0.25.0", features = ["extension-module"] }
 rayon = "1.10.0"
 serde_json = "1.0.119"
 serde = {version="1.0", features=["derive"]}
@@ -44,6 +44,7 @@ tempfile = "3.19.1"
 
 [build-dependencies]
 cbindgen = { version = "0.28.0", optional = true }
+pyo3-build-config = "0.25.0"
 
 [features]
 capi = ["libc"]
