### 2025-05-17 7806bfe
Use latest pixi version
