+++ .github/workflows/ci.yml
@@ -58,7 +58,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.8.8
         with:
-          pixi-version: v0.45.0
+          pixi-version: v0.47.0
           cache: true
           frozen: true
 
