### 2025-05-17 747f775
Use pixi for tests
