### 2025-05-18 e53e851
Fix pixi + MacOs test (#185)
* Add build file

* Add new pyo3 deps

* Use latest pixi version

* Use pixi for tests
