### 2025-06-03 #8187 @kthyng
@biavillasboas
* Is there a standard way to distribute rust packages you could use?
* I suggest something like readthedocs to compile and host your existing notebooks and docs if possible, unless I am missing something that already exists?
