### 2025-06-05 1fccde0
feat: Workflow to push static Python documentation
