+++ .github/workflows/python-doc.yml
@@ -1,22 +1,34 @@
 name: Python documentation
 
 on:
-  pull_request:
+  release:
+    types: [created, published]
   push:
     branches: [ "main" ]
   workflow_dispatch:
 
+permissions:
+  contents: read
+  pages: write
+  id-token: write
+
+concurrency:
+  group: "pages"
+  cancel-in-progress: false
+
 jobs:
-  build:
-    name: Build documentation
+  deploy:
+    environment:
+      name: github-pages
+      url: ${{ steps.deployment.outputs.page_url }}
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout sources
+      - name: Checkout
         uses: actions/checkout@v4
 
       - uses: prefix-dev/setup-pixi@v0.8.8
         with:
-          pixi-version: v0.45.0
+          pixi-version: v0.48.0
           cache: true
           frozen: true
 
@@ -27,3 +39,15 @@ jobs:
       - name: Build HTML
         # Turn warning into errors
         run: SPHINXOPTS="-W --keep-going -n" pixi run -e docs python-docs
+
+      - name: Setup Pages
+        uses: actions/configure-pages@v5
+
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@v3
+        with:
+          path: 'docs/_build/html'
+
+      - name: Deploy to GitHub Pages
+        id: deployment
+        uses: actions/deploy-pages@v4
