### 2025-06-05 52f6453
Publish static Python documentation (#188)
