+++ .github/workflows/python-test.yml
@@ -60,3 +60,24 @@ jobs:
           save-if: ${{ github.ref == 'refs/heads/main' }}
 
       - run: pixi run -e test pytest
+
+  doc:
+    name: Check documentation
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout sources
+        uses: actions/checkout@v4
+
+      - uses: prefix-dev/setup-pixi@v0.8.8
+        with:
+          pixi-version: v0.48.0
+          cache: true
+          frozen: true
+
+      - uses: Swatinem/rust-cache@v2
+        with:
+          save-if: ${{ github.ref == 'refs/heads/main' }}
+
+      - name: Build HTML
+        # Turn warning into errors
+        run: SPHINXOPTS="-W --keep-going -n" pixi run -e docs python-docs
