### 2025-06-05 #8187 @biavillasboas
@kthyng

Thank you for your comments.

- Rust packages are distributed through crates.io. Our package can be found here: https://crates.io/crates/mantaray
- We agree. We were all set to have the Python documentation on readthedocs, but then `mantaray` was taken by someone else in the meantime.  As an alternative, we decided to publish the documentation using [GitHub Pages](https://mines-oceanography.github.io/mantaray). The notebooks are now rendered as part of the documentation.
