### 2025-06-08 bb939d7
feat: Recording filenames and respective hashes
