### 2025-06-12 3241a17
Module `datasets` to provide example data (#189)
* feat, wip: datasets module to fetch example data using pooch

* fix: run ruff format on datasets.py

* feat: Recording filenames and respective hashes

* fix: Typo in one of the hashes

* feat: Functions to provide each dataset

* clean:

* cfg: Adding pooch as dependency

---------

Co-authored-by: Bryce Irving <bryceirving@mines.edu>
