+++ pyproject.toml
@@ -24,6 +24,8 @@ classifiers = [
 dependencies = [
   "cffi>=1.14.0",
   "xarray>=2023",
+  "pooch>=1.8.2",
+  "tqdm>=4.67.1",
   ]
 
 [project.optional-dependencies]
