### 2025-06-19 c46a847
Reconciling pixi.toml with pyproject.toml
Initially this was a Rust only project, using pixi.toml, but once we
added the Python wrapper, we also added the pyproject.toml. This
reconcile both in a single pyproject.toml.
