+++ .github/workflows/python-test.yml
@@ -49,7 +49,7 @@ jobs:
           fetch-depth: 0
 
       - name: Setup Pixi
-        uses: prefix-dev/setup-pixi@v0.8.8
+        uses: prefix-dev/setup-pixi@v0.8.14
         with:
           pixi-version: v0.45.0
           frozen: true
@@ -68,7 +68,7 @@ jobs:
       - name: Checkout sources
         uses: actions/checkout@v4
 
-      - uses: prefix-dev/setup-pixi@v0.8.8
+      - uses: prefix-dev/setup-pixi@v0.8.14
         with:
           pixi-version: v0.48.0
           cache: true
