### 2025-07-27 7ff3bb1
cfg: Updating pixi used in GA
