+++ .github/workflows/python-test.yml
@@ -51,7 +51,7 @@ jobs:
       - name: Setup Pixi
         uses: prefix-dev/setup-pixi@v0.8.14
         with:
-          pixi-version: v0.45.0
+          pixi-version: v0.50.1
           frozen: true
           cache: true
 
@@ -70,7 +70,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.8.14
         with:
-          pixi-version: v0.48.0
+          pixi-version: v0.50.1
           cache: true
           frozen: true
 
