### 2025-07-27 728aee5
cfg: With pixi, use Python at least 3.12
