+++ pyproject.toml
@@ -62,7 +62,7 @@ python-docs = { cmd = "make html", cwd = "docs", depends-on = ["develop"] }
 [tool.pixi.dependencies]
 maturin = ">=1.5.1,<1.6"
 rust = ">=1.84.0,<1.85"
-python = ">=3.11"
+python = ">=3.12"
 cffi = ">=1.16.0,<1.17"
 pip = ">=24.0,<25"
 xarray = ">=2023"
