### 2025-07-27 48919ee
cfg: Updating ruff
