+++ pyproject.toml
@@ -67,7 +67,7 @@ cffi = ">=1.16.0,<1.17"
 pip = ">=24.0,<25"
 xarray = ">=2023"
 netcdf4 = ">=1.6.5,<1.7"
-ruff = ">=0.4.9,<0.5"
+ruff = ">=0.12.5,<0.13"
 matplotlib = ">=3.9.1,<3.10"
 cmake = ">=3.30.5,<4"
 pooch = ">=1.8.2,<2"
