+++ pyproject.toml
@@ -66,7 +66,7 @@ python = ">=3.12"
 cffi = ">=1.16.0,<1.17"
 pip = ">=24.0,<25"
 xarray = ">=2023"
-netcdf4 = ">=1.6.5,<1.7"
+netcdf4 = ">=1.7.2,<2"
 ruff = ">=0.12.5,<0.13"
 matplotlib = ">=3.9.1,<3.10"
 cmake = ">=3.30.5,<4"
@@ -102,7 +102,7 @@ myst-sphinx-gallery = ">=0.3.0,<0.4"
 
 [tool.pixi.feature.test.dependencies]
 pytest = ">=7.0"
-netcdf4 = ">=1.5"
+netcdf4 = ">=1.7.2,<2"
 
 [tool.pixi.feature.examples.dependencies]
 jupyterlab = "*"
