+++ Cargo.toml
@@ -26,7 +26,7 @@ crate-type = ["cdylib", "lib"]
 
 [dependencies]
 derive_builder = "0.20.0"
-libc = { version = "0.2.172", optional = true }
+libc = { version = "0.2.174", optional = true }
 ndarray = { version = "0.16.1", features = ["rayon"] }
 netcdf3 = "0.5.2"
 netcdf = {version = "0.11.0", features = ["static", "ndarray"]}
