+++ Cargo.toml
@@ -40,7 +40,7 @@ tracing = "0.1.41"
 
 [dev-dependencies]
 lockfile = "0.4.0"
-tempfile = "3.19.1"
+tempfile = "3.20.0"
 
 [build-dependencies]
 cbindgen = { version = "0.28.0", optional = true }
