### 2025-07-27 d6df67b
cfg: Updating serde_json
