+++ Cargo.toml
@@ -33,7 +33,7 @@ netcdf = {version = "0.11.0", features = ["static", "ndarray"]}
 ode_solvers = "0.4.0"
 pyo3 = { version = "0.25.0", features = ["extension-module"] }
 rayon = "1.10.0"
-serde_json = "1.0.119"
+serde_json = "1.0.141"
 serde = {version="1.0", features=["derive"]}
 thiserror = "2.0.12"
 tracing = "0.1.41"
