### 2025-07-27 778cb0f
cfg: Adding uv to dev environment
