### 2025-07-27 7daf599
cfg: Ruff is only required on dev
