+++ pyproject.toml
@@ -67,7 +67,6 @@ cffi = ">=1.16.0,<1.17"
 pip = ">=24.0,<25"
 xarray = ">=2023"
 netcdf4 = ">=1.7.2,<2"
-ruff = ">=0.12.5,<0.13"
 matplotlib = ">=3.9.1,<3.10"
 cmake = ">=3.30.5,<4"
 pooch = ">=1.8.2,<2"
@@ -84,7 +83,7 @@ test = { features = ["test"], solve-group = "default" }
 examples = { features = ["dev", "docs", "examples"], solve-group = "default" }
 
 [tool.pixi.feature.dev.dependencies]
-ruff = ">=0.4"
+ruff = ">=0.12.5,<0.13"
 ipython = ">=8.0"
 uv = ">=0.8.3,<0.9"
 
