### 2025-07-27 d9db829
cfg: Replacing build by uv for development (release)
