+++ pyproject.toml
@@ -31,7 +31,7 @@ dependencies = [
 [project.optional-dependencies]
 netcdf = ["netcdf4>=1.5.7"]
 dev = [
-  "build>=0.6",
+  "uv>=0.8",
 ]
 test = [
   "pytest>=7.0",
