### 2025-07-27 8def07e
Updating some dependencies (#193)
* cfg: Updating pixi used in GA

* cfg: Updating pixi.lock

* cfg: With pixi, use Python at least 3.12

* cfg: Updating ruff

* cfg: Updating netCDF4

* cfg: Updating libc

* cfg: Updating tempfile

* cfg: Updating serde_json

* cfg: Updating pyo3

* cfg: Updating Cargo.lock

* cfg: Adding uv to dev environment

* cfg: Ruff is only required on dev

* cfg: Replacing build by uv for development (release)
