+++ Cargo.toml
@@ -26,21 +26,21 @@ crate-type = ["cdylib", "lib"]
 
 [dependencies]
 derive_builder = "0.20.0"
-libc = { version = "0.2.172", optional = true }
+libc = { version = "0.2.174", optional = true }
 ndarray = { version = "0.16.1", features = ["rayon"] }
 netcdf3 = "0.5.2"
 netcdf = {version = "0.11.0", features = ["static", "ndarray"]}
 ode_solvers = "0.4.0"
-pyo3 = { version = "0.25.0", features = ["extension-module"] }
+pyo3 = { version = "0.25.1", features = ["extension-module"] }
 rayon = "1.10.0"
-serde_json = "1.0.119"
+serde_json = "1.0.141"
 serde = {version="1.0", features=["derive"]}
 thiserror = "2.0.12"
 tracing = "0.1.41"
 
 [dev-dependencies]
 lockfile = "0.4.0"
-tempfile = "3.19.1"
+tempfile = "3.20.0"
 
 [build-dependencies]
 cbindgen = { version = "0.28.0", optional = true }
