+++ pyproject.toml
@@ -31,7 +31,7 @@ dependencies = [
 [project.optional-dependencies]
 netcdf = ["netcdf4>=1.5.7"]
 dev = [
-  "build>=0.6",
+  "uv>=0.8",
 ]
 test = [
   "pytest>=7.0",
@@ -62,12 +62,11 @@ python-docs = { cmd = "make html", cwd = "docs", depends-on = ["develop"] }
 [tool.pixi.dependencies]
 maturin = ">=1.5.1,<1.6"
 rust = ">=1.84.0,<1.85"
-python = ">=3.11"
+python = ">=3.12"
 cffi = ">=1.16.0,<1.17"
 pip = ">=24.0,<25"
 xarray = ">=2023"
-netcdf4 = ">=1.6.5,<1.7"
-ruff = ">=0.4.9,<0.5"
+netcdf4 = ">=1.7.2,<2"
 matplotlib = ">=3.9.1,<3.10"
 cmake = ">=3.30.5,<4"
 pooch = ">=1.8.2,<2"
@@ -84,8 +83,9 @@ test = { features = ["test"], solve-group = "default" }
 examples = { features = ["dev", "docs", "examples"], solve-group = "default" }
 
 [tool.pixi.feature.dev.dependencies]
-ruff = ">=0.4"
+ruff = ">=0.12.5,<0.13"
 ipython = ">=8.0"
+uv = ">=0.8.3,<0.9"
 
 [tool.pixi.feature.docs.dependencies]
 furo = ">=2024.8.6,<2025"
@@ -102,7 +102,7 @@ myst-sphinx-gallery = ">=0.3.0,<0.4"
 
 [tool.pixi.feature.test.dependencies]
 pytest = ">=7.0"
-netcdf4 = ">=1.5"
+netcdf4 = ">=1.7.2,<2"
 
 [tool.pixi.feature.examples.dependencies]
 jupyterlab = "*"
