+++ .github/workflows/ci.yml
@@ -17,7 +17,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.8.14
         with:
-          pixi-version: v0.45.0
+          pixi-version: v0.50.1
           cache: true
           frozen: true
 
@@ -58,7 +58,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.8.14
         with:
-          pixi-version: v0.47.0
+          pixi-version: v0.50.1
           cache: true
           frozen: true
 
@@ -79,7 +79,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.8.14
         with:
-          pixi-version: v0.45.0
+          pixi-version: v0.50.1
           cache: true
           frozen: true
 
