+++ .github/workflows/python-doc.yml
@@ -28,7 +28,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.8.14
         with:
-          pixi-version: v0.48.0
+          pixi-version: v0.50.1
           cache: true
           frozen: true
 
