### 2025-07-28 4231d63
refact: Moving LinearFit into its own sub-module
