### 2025-07-31 #8187 @milancurcic
Yes, very happy to.
