### 2025-08-01 #8691 @Sbozzolo
@milancurcic, @gauteh, thank you so much for agreeing to review!

Please, go ahead, generate your checklist, and have a look at `mantaray`.

I'll be there for any questions you might have.

Thanks again!
