### 2025-08-06 #8691 @gauteh
In equation for dx/dt it may be useful to specify that x with dot above is time derivative or vector? Similar for wave number vector.
