### 2025-08-06 #8691 @gauteh
I've gone through the paper, cloned and installed the software (using `pip`), and run the various notebooks. Everything went very smooth indeed (although I am used to dealing with mixed rust+python projects), so I think there are good chances this will also go smooth for future non-rusty users!

Your examples are very useful, and are the most important part of the documentation. Can't wait to test this package on some of our own projects! I think improving the documentation would benefit the project and help future users, but not much is required to satisfy minimum requirements. I'll wait for your comments before ticking the final boxes.
