### 2025-08-10 b58ffb7
fix: remove unused lockfile dependency (#195)
