+++ Cargo.toml
@@ -39,7 +39,6 @@ thiserror = "2.0.12"
 tracing = "0.1.41"
 
 [dev-dependencies]
-lockfile = "0.4.0"
 tempfile = "3.20.0"
 
 [build-dependencies]
