### 2025-08-11 mines-oceanography/mantaray#198 @milancurcic
Add instructions for installing notebook packages
There's a nice notebook gallery at https://mines-oceanography.github.io/mantaray/gallery_header.html# and I'd like to run the examples. However, the notebooks don't seem to provide instructions on how to actually run the notebook. Jupyter is not installed with the package (good!) and I'm new to pixi so I had to do some digging on how to add Jupyter and related packages I use using:

```
pixi add jupyter ipython matplotlib
```

However, the top-level notebook gallery should provide instructions on installing notebook tools via pixi.

(part of https://github.com/openjournals/joss-reviews/issues/8691)
