### 2025-08-11 mines-oceanography/mantaray#199 @milancurcic
`cmocean` is a required dependency but is not installed
Canonical examples notebook docs at https://mines-oceanography.github.io/mantaray/canonical_ray_tracing.html instruct importing `cmocean` but `cmocean` is not a core dependency:

```python
In [2]: # data handling
   ...: import numpy as np
   ...: import xarray as xr
   ...:
   ...: # plotting
   ...: import matplotlib.pyplot as plt
   ...: from IPython.display import HTML
   ...: import cmocean
   ...:
   ...: # ray tracing
   ...: import mantaray
   ...:
   ...: # helper functions
   ...: import utils
Cell In[2], line 8
      6 import matplotlib.pyplot as plt
      7 from IPython.display import HTML
----> 8 import cmocean
     10 # ray tracing
     11 import mantaray

ModuleNotFoundError: No module named 'cmocean'
```

Libraries used in the example notebooks should either be included by default, or, more preferably IMO, should provide instructions on installing these optional dependencies.

Similar for `utils`. I see that it's included in notebooks/canonical_examples/utils.py, but this is not documented. I think each notebook is intended to be run from its respective directory. If so, this should also be part of the notebook instructions.

Part of https://github.com/openjournals/joss-reviews/issues/8691
