### 2025-08-11 mines-oceanography/mantaray#200 @milancurcic
Place the "Constructing canonical forcing fields" notebook before the "Canonical examples" notebook
I appreciated this note in the canonical examples notebook:

> Introduction
> This notebook covers how to run Mantaray with the custom depth and current fields that were constructed in idealized_fields.ipynb. It is recommended that users run through this notebook first.

Here it's stated as optional, so I proceeded with the canonical examples. However, the canonical examples notebook requires this dataset:

```python
ds_jet = xr.open_dataset("./forcing/zonal_jet.nc")
```

which, after looking inside forcing/ and the README there, I understood that I actually need to run idealized_fields.ipynb first to generate the data.

This is fine, but it can and should be much simplified by:

* Providing idealized_fields.ipynb notebook in the Gallery and have it appear before the canonical examples
* Stating that this is required, e.g.:

> Introduction
> This notebook covers how to run Mantaray with the custom depth and current fields that were constructed in idealized_fields.ipynb. It is required that users run through this notebook first.

Part of https://github.com/openjournals/joss-reviews/issues/8691
