### 2025-08-14 mines-oceanography/mantaray#200 @milancurcic
After finishing the Canonical examples, I notice that the next notebook is "Constructing canonical fields" so this issue is already solved! :) I changed the issue title to specifically ask for reordering of notebooks.
