### 2025-08-14 #8691 @milancurcic
I am now finished with the review. The only aspects needing improvement are installation instructions (Jupyter notebooks are provided but not the instructions to install and run Jupyter itself; same for the optional `cmocean` package which is used by the notebooks). The notebooks are also a bit rough around the edges; a clean-up and reordering (see https://github.com/mines-oceanography/mantaray/issues/200) would be welcome.

Some minor recommendations for the text:

* L27: "written in low-level languages such as Fortran or C/C++": I understand that here you mean it compared to Python, but then the numerical implementation here is in Rust, which for a physics solver is arguably lower-level than Fortran. :). Fortran, C, C++, and Rust can all be integrated with a high-level Python wrapper, as it's done here, so the argument for Rust is not so clear. May I suggest an alternative motivation that an implementation in a newer language such as Rust is good for both the wave physics and the programming communities by a diversity argument. Rust is clearly showing capability and readiness for physics solvers like this and it's great to have this package.
* It seems from the repo (Rust packaging and Jupyter notebooks) that there is a limitation on the NetCDF-3 classic format. When I tried to load a NetCDF4 (HDF5 under the hood), the function failed, I believe at the Rust level. I suspect that the limitation to NetCDF-3 classic format is simply a Rust library ecosystem maturity deficiency and will be resolved over time. Nevertheless, it would be good to point this limitation out in the paper, if it is indeed a limitation and not just my misuse of the library.
