### 2025-08-20 10d1e9b
Change citations to square brackets
