### 2025-08-20 eb15961
Add definition of time derivative
