### 2025-08-20 20b436d
Address @milancurcic comment regarding NetCDF3
