### 2025-08-21 272076d
cfg: Only on new versions
