+++ .github/workflows/release-python.yml
@@ -5,7 +5,7 @@ on:
     branches:
       - main
     tags:
-      - '*'
+      - 'v*.*.*'
   pull_request:
   workflow_dispatch:
 
