### 2025-08-21 8b56071
fix: Dynamic versioning requires the tags history
