+++ .github/workflows/release-python.yml
@@ -32,6 +32,9 @@ jobs:
           #   target: ppc64le
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+          fetch-tags: true
       - uses: actions/setup-python@v5
         with:
           python-version: '3.10'
@@ -59,6 +62,9 @@ jobs:
             target: x86
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+          fetch-tags: true
       - uses: actions/setup-python@v5
         with:
           python-version: '3.10'
@@ -86,6 +92,9 @@ jobs:
             target: aarch64
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+          fetch-tags: true
       - uses: actions/setup-python@v5
         with:
           python-version: '3.10'
@@ -105,6 +114,9 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+          fetch-tags: true
       - name: Build sdist
         uses: PyO3/maturin-action@v1
         with:
