### 2025-08-21 dcca9ca
cfg: Removing manual dispatch
