+++ .github/workflows/release-python.yml
@@ -7,7 +7,6 @@ on:
     tags:
       - 'v*.*.*'
   pull_request:
-  workflow_dispatch:
 
 permissions:
   contents: read
