+++ .github/workflows/release-python.yml
@@ -6,7 +6,6 @@ on:
       - main
     tags:
       - 'v*.*.*'
-  pull_request:
 
 permissions:
   contents: read
