### 2025-08-21 04b4726
Fix latex typo
