### 2025-08-21 3586457
Reverting notation to parenthesis to test pdf compilation
