### 2025-08-26 6b23f55
Fix version submitted to pypi (#203)
* fix: Dynamic versioning requires the tags history

* cfg: Removing manual dispatch

* cfg: Remove PR triggered
