+++ .github/workflows/python-doc.yml
@@ -24,7 +24,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
 
       - uses: prefix-dev/setup-pixi@v0.8.14
         with:
