### 2025-08-29 abedb91
update dependency `ode_solvers` (#197)
