+++ Cargo.toml
@@ -30,7 +30,7 @@ libc = { version = "0.2.174", optional = true }
 ndarray = { version = "0.16.1", features = ["rayon"] }
 netcdf3 = "0.5.2"
 netcdf = {version = "0.11.0", features = ["static", "ndarray"]}
-ode_solvers = "0.4.0"
+ode_solvers = "0.6.1"
 pyo3 = { version = "0.25.1", features = ["extension-module"] }
 rayon = "1.10.0"
 serde_json = "1.0.141"
