+++ .github/workflows/python-test.yml
@@ -49,7 +49,7 @@ jobs:
           fetch-depth: 0
 
       - name: Setup Pixi
-        uses: prefix-dev/setup-pixi@v0.8.14
+        uses: prefix-dev/setup-pixi@v0.9.0
         with:
           pixi-version: v0.50.1
           frozen: true
@@ -68,7 +68,7 @@ jobs:
       - name: Checkout sources
         uses: actions/checkout@v5
 
-      - uses: prefix-dev/setup-pixi@v0.8.14
+      - uses: prefix-dev/setup-pixi@v0.9.0
         with:
           pixi-version: v0.50.1
           cache: true
