+++ .github/workflows/ci.yml
@@ -15,7 +15,7 @@ jobs:
       - name: Checkout sources
         uses: actions/checkout@v5
 
-      - uses: prefix-dev/setup-pixi@v0.8.14
+      - uses: prefix-dev/setup-pixi@v0.9.0
         with:
           pixi-version: v0.50.1
           cache: true
@@ -56,7 +56,7 @@ jobs:
           toolchain: ${{ matrix.rust }}
           override: true
 
-      - uses: prefix-dev/setup-pixi@v0.8.14
+      - uses: prefix-dev/setup-pixi@v0.9.0
         with:
           pixi-version: v0.50.1
           cache: true
@@ -77,7 +77,7 @@ jobs:
       - name: Checkout sources
         uses: actions/checkout@v5
 
-      - uses: prefix-dev/setup-pixi@v0.8.14
+      - uses: prefix-dev/setup-pixi@v0.9.0
         with:
           pixi-version: v0.50.1
           cache: true
