### 2025-08-29 #8691 @biavillasboas
Hi, @Sbozzolo.
Sorry for the slow reply and thank you for the nudge! Reviews are going well, and we've worked through most of the comments (which haven't been merged to main yet). This has been a bit slow because I'd like Bryce, who is the undergrad who led the work, to take part in the review process as a learning experience. Now that the semester has started and he's back, we should be able to wrap up pretty quickly. Thanks!
