### 2025-09-03 3ccd519
Addresses white space issue in GIF (#172)
* Add templates for Jupyter notebook examples

* work in progress commit

* changes forcing files to have more realistic scales

* updates notebooks to change scales of domain

* removes bloat and uneeded code

* ingores data files

* ingores data files

* ingores data files

* removes testing data files

* removes data files from tracking

* adds data directort to .gitignore

* changes code to not depend on scipy

* removes testing dataset

* changes path for data files

* moves notebooks to notebooks folder

* removes old example notebook

* creates 
[…]
