### 2025-09-03 4e06167
Replace comma by semicolon for separating references
