+++ joss/paper.md
@@ -45,7 +45,7 @@ Ocean surface gravity waves are an important component of air-sea interaction, i
 ![Examples of ray tracing performed using `Mantaray`. Top left: waves in deep water interacting with a zonal jet. Bottom left: waves in deep water interacting with a mesoscale eddy. Top right: waves encountaring varyring bathymethy approaching a linear beach. Bottom right: waves approaching a Gaussian island. \label{fig:examples}](idealized_showcase.png){ width=100% }
 
 # Statement of need
-Ray tracing is a long-standing method for investigating wave propagation across a wide range of disciplines, including optics, seismology, and oceanography, providing a simple framework for studying the evolution of waves in spatially varying media. For ocean surface gravity waves, ray-based approaches have been used to study refraction by mesoscale currents (e.g., @mapp1985wave, @romero2017observations, @marechal2022variability), changes in bathymetry [e.g., @munk1947refraction, @kukulka2017surface], and statistical effects such as directional diffusion of wave action [e.g., @smit2019swell, @VBY2023]. 
+Ray tracing is a long-standing method for investigating wave propagation across a wide range of disciplines, including optics, seismology, and oceanography, providing a simple framework for studying the evolution of waves in spatially varying media. For ocean surface gravity waves, ray-based approaches have been used to study refraction by mesoscale currents [e.g., @mapp1985wave; @romero2017observations; @marechal2022variability], changes in bathymetry [e.g., @munk1947refraction; @kukulka2017surface], and statistical effects such as directional diffusion of wave action [e.g., @smit2019swell; @VBY2023]. 
 
 Although ray tracing has been widely used in surface wave studies, the software implementations are often written in Fortran or C/C++ [e.g., @oreilly2016california] and are not always openly available or actively maintained. More recently, open-source Python tools—such as the one by @halsne2023ocean—have improved accessibility and reproducibility. Mantaray complements these efforts by providing a ray tracing solution built in Rust, a modern  programming language that combines memory safety and execution speed with tools for seamless Python integration. This choice balances the ease-of-use associated with Python and the computational efficiency of compiled languages, filling a gap for users who need robust, high-performance ray tracing within a user-friendly environment.
 
