### 2025-09-03 mines-oceanography/mantaray#198 @milancurcic
I think these few clarifications is all that's needed.
