### 2025-09-07 d7e4fe5
Addressed editorial comments from the reviewers (#208)
* Change citations to square brackets

* Add definition of time derivative

* Adress @milancurcic comment regarding the value of the package

* Address @milancurcic comment regarding NetCDF3

* Fix latex typo

* Reverting notation to parenthesis to test pdf compilation

* Replace comma by semicolon for separating references
