### 2025-09-07 5a0c070
cfg: Adding project's optional 'examples'
Add extra requirements to run the examples.
