+++ pyproject.toml
@@ -29,6 +29,10 @@ dependencies = [
   ]
 
 [project.optional-dependencies]
+examples = [
+  "jupyterlab>=4.4.5",
+  "cmocean>=4.0.3",
+]
 netcdf = ["netcdf4>=1.5.7"]
 dev = [
   "uv>=0.8",
