### 2025-09-07 759f4f4
cfg: Defining minimum jupyterlab version for pixi
