+++ pyproject.toml
@@ -30,8 +30,8 @@ dependencies = [
 
 [project.optional-dependencies]
 examples = [
-  "jupyterlab>=4.4.5",
   "cmocean>=4.0.3",
+  "jupyterlab>=4.4.5",
 ]
 netcdf = ["netcdf4>=1.5.7"]
 dev = [
