### 2025-09-07 c98c5a4
cfg: Updating maturin used in pixi
