+++ pyproject.toml
@@ -60,7 +60,7 @@ python-docs = { cmd = "make html", cwd = "docs", depends-on = ["develop"] }
 
 
 [tool.pixi.dependencies]
-maturin = ">=1.5.1,<1.6"
+maturin = ">=1.9.4,<2"
 rust = ">=1.84.0,<1.85"
 python = ">=3.12"
 cffi = ">=1.16.0,<1.17"
