### 2025-09-07 84b077f
cfg: Manually bump Python package version
