### 2025-09-07 0b212c2
fix PyPI version release (#213)
* cfg: Updating maturin used in pixi

* cfg: Manually bump Python package version
