### 2025-09-08 9a6dadf
fix: update nazare example docs to include data link
