+++ notebooks/nazare/README.md
@@ -1,0 +1,6 @@
+# Ray Tracing in the Nazare Canyon
+
+The `notebook/` subdirectory includes the `nazare_example.ipynb` jupyter notebook and the supporting functions in `src.py`.
+
+## Additional requirements
+Before running this example, create a folder `data/` within this directory and download and place the required data files from the [sample data release](https://github.com/mines-oceanography/mantaray/releases/tag/d0.0.7) in there. For this example, you will need to download `spec_s2.nc`, `s2_image.nc`, and `nazare_bathy.nc`.
