### 2025-09-10 b22488f
fix: update example docs
