### 2025-09-10 mines-oceanography/mantaray#198 @brycemines
> Let me know if I should open another issue for this, but given what @castelao described in https://github.com/mines-oceanography/mantaray/issues/204, I dont think it would be much more work for me to add a requirements.txt to each example folder.

Leaving this for another issue: #217
