### 2025-09-14 f5af909
Update example documentation and instructions (#216)
* update mantaray readme

* fix: update example docs
