### 2025-09-29 2631dee
Adding disclaimer required by NREL (#222)
