### 2025-09-30 3ce00fd
Add badges to the README
This commit adds badges that link to the documentation as well as Rust and Python package distributions.
