### 2025-10-12 32bb5d4
fix: Remove unused html tags
