+++ docs/index.md
@@ -5,8 +5,6 @@ hide-toc: true
 # Welcome to Mantaray's documentation!
 
 ```{include} ../README.md
-:start-after: <!-- start elevator-pitch -->
-:end-before: <!-- end elevator-pitch -->
 ```
 
 ```{toctree}
