### 2025-10-13 c11000a
cfg, fix: Missing some dependencies
Thanks to @brycemines who noticed some missing dependencies.
