+++ pyproject.toml
@@ -32,6 +32,8 @@ dependencies = [
 examples = [
   "cmocean>=4.0.3",
   "jupyterlab>=4.4.5",
+  "netcdf4>=1.5.7",
+  "matplotlib>=3.9.1",
 ]
 netcdf = ["netcdf4>=1.5.7"]
 dev = [
