### 2025-10-16 7ffe0a9
fix: update the example docs
