### 2025-10-22 af17040
Optional examples (#212)
* cfg: Adding project's optional 'examples'

Add extra requirements to run the examples.

* cfg: Defining minimum jupyterlab version for pixi

* cfg, style: Alphabetical order

* cfg, fix: Missing some dependencies

Thanks to @brycemines who noticed some missing dependencies.
