+++ pyproject.toml
@@ -29,6 +29,12 @@ dependencies = [
   ]
 
 [project.optional-dependencies]
+examples = [
+  "cmocean>=4.0.3",
+  "jupyterlab>=4.4.5",
+  "netcdf4>=1.5.7",
+  "matplotlib>=3.9.1",
+]
 netcdf = ["netcdf4>=1.5.7"]
 dev = [
   "uv>=0.8",
@@ -105,7 +111,7 @@ pytest = ">=7.0"
 netcdf4 = ">=1.7.2,<2"
 
 [tool.pixi.feature.examples.dependencies]
-jupyterlab = "*"
+jupyterlab = ">=4.4.7,<5"
 cmocean = ">=4.0.3,<5"
 
 [tool.maturin]
