### 2025-10-22 87248f8
cfg: New pixi standard - workspace
