+++ pyproject.toml
@@ -50,7 +50,7 @@ test = [
 #build = "cp310-*"
 
 
-[tool.pixi.project]
+[tool.pixi.workspace]
 channels = ["conda-forge"]
 platforms = ["osx-arm64", "linux-64", "win-64"]
 
