### 2025-10-22 a10ef8e
cfg: Updating pixi version used at GA
