+++ .github/workflows/python-doc.yml
@@ -28,7 +28,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.9.2
         with:
-          pixi-version: v0.50.1
+          pixi-version: v0.57.0
           cache: true
           frozen: true
 
