### 2025-10-22 f6ab3ef
update nazare and agulhas examples to use mantaray datasets (#215)
* fix: update nazare example docs to include data link

* refactor nazare example to use mantaray.datasets

* delete the work in progress line for mantaray datasets

* refactor agulhas example to use datasets module

* fix: update the example docs

---------

Co-authored-by: Bia Villas Boas <villasboas@mines.edu>
