+++ notebooks/nazare/README.md
@@ -1,0 +1,3 @@
+# Ray Tracing in the Nazare Canyon
+
+The `notebook/` subdirectory includes the `nazare_example.ipynb` jupyter notebook and the supporting functions in `src.py`.
