### 2025-10-22 20f912e
Regular update (#229)
* cfg: New pixi standard - workspace

* cfg: Updating pixi.lock

* cfg: Updating pixi version used at GA

---------

Co-authored-by: Bia Villas Boas <villasboas@mines.edu>
