+++ .github/workflows/python-test.yml
@@ -51,7 +51,7 @@ jobs:
       - name: Setup Pixi
         uses: prefix-dev/setup-pixi@v0.9.2
         with:
-          pixi-version: v0.50.1
+          pixi-version: v0.57.0
           frozen: true
           cache: true
 
@@ -70,7 +70,7 @@ jobs:
 
       - uses: prefix-dev/setup-pixi@v0.9.2
         with:
-          pixi-version: v0.50.1
+          pixi-version: v0.57.0
           cache: true
           frozen: true
 
