### 2025-10-26 56e8253
Creating credits
