+++ AUTHORS.md
@@ -1,0 +1,6 @@
+# Credits
+
+## Development Lead
+
+* Bryce Irving <bryceirving@mines.edu>
+* Guilherme Castelão <guilherme@castelao.net>
