+++ AUTHORS.md
@@ -4,3 +4,21 @@
 
 * Bryce Irving <bryceirving@mines.edu>
 * Guilherme Castelão <guilherme@castelao.net>
+* Bia Villas Boas <villasboas@mines.edu>
+
+## Development Support
+
+* @npizzo
+* @colinbeyers
+* @jmclemson
+* @jackson-krieger
+* @gmarechal
+
+## Contributors
+
+* @luizirber: Several software design recommendations.
+* @ppinchukk: PR #185 - Improved PYO3 setup.
+* @iuryt: Tested and reviwed earlier versions of the package.
+* @milancurcic: Several recommendations in the JOSS review process.
+* @gauteh: Several recommendations in the JOSS review process.
+* @Sbozzolo: Several recommendations in the JOSS review process.
