### 2025-10-26 c298e78
Defining contributors (#230)
* Initializing .zenodo.json

* Adding contributors

* Creating credits

* Listing contributors

* Fixing Bia's name order

* Adding (software) references
