### 2025-10-26 7ab21b7
Copy animation
