### 2025-10-26 d954ec0
Fix broken image in landing page (#232)
* cfg: Local link to demo

Skip root README and define its own link to demo.

* Copy animation

* cfg: Re-organizing pixi's environments

* Updating pixi.lock
