+++ docs/index.md
@@ -4,7 +4,13 @@ hide-toc: true
 
 # Welcome to Mantaray's documentation!
 
+A library for surface gravity waves ray tracing.
+
+![Demo](demo_animation.gif)
+
 ```{include} ../README.md
+:parser: myst_parser.sphinx_
+:start-after: <!-- examples -->
 ```
 
 ```{toctree}
