+++ .github/workflows/release-python.yml
@@ -132,7 +132,7 @@ jobs:
     if: "startsWith(github.ref, 'refs/tags/')"
     needs: [linux, windows, macos, sdist]
     steps:
-      - uses: actions/download-artifact@v5
+      - uses: actions/download-artifact@v6
       - name: Publish to PyPI
         uses: PyO3/maturin-action@v1
         env:
