### 2025-10-27 cd9e50a
fix: remove duplicate link
