### 2025-11-05 8b2ea34
fix: update readme, users cant add labels (#236)
