### 2025-11-05 71f2155
Fix typo in the python examples
This commit fixes issue #239. Some parts of the code example shown in the paper used phi0 as the initial direction, when our convention is theta0.
