### 2025-11-05 0ef5d22
Revise acknowledgements in paper.md
Updated acknowledgements section to include additional support and reviewers. Fixes issue #231
