### 2025-11-05 1c92e26
Updates dispersion relationship
Updates the convention in the dispersion relationship equation to comply with NREL's internal review.
