### 2025-11-12 0e493f5
Add DOI information to bibliography entries
Added DOI entries for various articles in the bibliography.
