+++ joss/paper.bib
@@ -5,7 +5,8 @@
   volume={890},
   pages={R3},
   year={2020},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.1017/jfm.2020.116}
 }
 
 @article{smit2019swell,
@@ -15,7 +16,8 @@
   volume={49},
   number={10},
   pages={2615--2630},
-  year={2019}
+  year={2019},
+  doi={10.1175/JPO-D-18-0250.1}
 }
 
 @article{halsne2023ocean,
@@ -26,7 +28,8 @@
   number={22},
   pages={6515--6530},
   year={2023},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+  publisher={Copernicus Publications G{\"o}ttingen, Germany},
+  doi={10.5194/gmd-16-6515-2023}
 }
 
 @article{oreilly2016california,
@@ -36,7 +39,8 @@
   volume={116},
   pages={118--132},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.coastaleng.2016.06.005}
 }
 
 @article{mapp1985wave,
@@ -47,7 +51,8 @@
   number={C4},
   pages={7153--7162},
   year={1985},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/JC090iC04p07153}
 }
 
 @article{romero2017observations,
@@ -57,7 +62,8 @@
   volume={47},
   number={3},
   pages={615--632},
-  year={2017}
+  year={2017},
+  doi={10.1175/JPO-D-16-0108.1}
 }
 
 @article{marechal2022variability,
@@ -68,7 +74,8 @@
   number={5},
   pages={1275--1292},
   year={2022},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+  publisher={Copernicus Publications G{\"o}ttingen, Germany},
+  doi={10.5194/os-18-1275-2022}
 }
 
 @article{munk1947refraction,
@@ -79,7 +86,8 @@
   number={1},
   pages={1--26},
   year={1947},
-  publisher={University of Chicago Press}
+  publisher={University of Chicago Press},
+  doi={10.1086/625388}
 }
 
 @article{kukulka2017surface,
@@ -90,14 +98,15 @@
   number={11},
   pages={8683--8706},
   year={2017},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/2017JC013370}
 }
 
 @book{phillips1966,
-	title = {The {Dynamics} of the {Upper} {Ocean}},
-	language = {en},
-	urldate = {2024-11-14},
-	publisher = {Cambridge University Press},
-	author = {Phillips, Owen. M},
-	year = {1966},
+  title = {The {Dynamics} of the {Upper} {Ocean}},
+  language = {en},
+  urldate = {2024-11-14},
+  publisher = {Cambridge University Press},
+  author = {Phillips, Owen. M},
+  year = {1966},
 }
