### 2025-11-12 #8691 @biavillasboas
@castelao, I'm addressing all comments, except for

> * For development, you suggest `git clone git@github.com:mines-oceanography/mantaray.git`. This requires a GitHub account (because you are using ssh instead of https) and won't allow external contributors to push to the repo (because they won't have permissions). Somewhere else you suggest forking the repo. Maybe you want to clarify a tiny bit your recommendations.

Could you take a look at that one when you have a chance?
