### 2025-11-24 8f7b630
Update acknowledgements to include data sources
Added acknowledgment of data sources used in example notebooks.
