+++ joss/paper.md
@@ -138,6 +138,6 @@ ray_path = mantaray.ray_tracing(x0, y0, kx0, ky0,
 
 # Acknowledgements
 This work was authored in part by NREL for the U.S. Department of Energy (DOE), operated under Contract No. DE-AC36-08GO28308.
-We thank the Colorado School of Mines Summer Undergraduate Research Fellowship (SURF) and the Mines Undergraduate Research Fellowship (MURF) for partially supporting undergraduate students BI, JC, and JK. BVB was supported by the ONR award N00014-24-1-2554 through the SASCWATCH MURI, and NASA award 80NSSC24K1640 through the SWOT Science Team. CB was supported by NASA awards 80NSSC23K0979 through the International Ocean Vector Winds Science Team and 80GSFC24CA067 through the ODYSEA science team as part of the Earth System Explorer program. JK and GM were supported by NASA award 80NSSC24K0411 through the S-MODE Science Team. All co-authors are thankful to Luiz Irber for helpful recommendations throughout the development of this package. The authors also thank Milan Curcic and Gaute Hope for reviewing this paper, and Gabriele Bozzola for serving as editor. The views expressed in the article do not necessarily represent the views of the DOE or the U.S. Government.
+We thank the Colorado School of Mines Summer Undergraduate Research Fellowship (SURF) and the Mines Undergraduate Research Fellowship (MURF) for partially supporting undergraduate students BI, JC, and JK. BVB was supported by the ONR award N00014-24-1-2554 through the SASCWATCH MURI, and NASA award 80NSSC24K1640 through the SWOT Science Team. CB was supported by NASA awards 80NSSC23K0979 through the International Ocean Vector Winds Science Team and 80GSFC24CA067 through the ODYSEA science team as part of the Earth System Explorer program. JK and GM were supported by NASA award 80NSSC24K0411 through the S-MODE Science Team. All co-authors are thankful to Luiz Irber for helpful recommendations throughout the development of this package. The authors also thank Milan Curcic and Gaute Hope for reviewing this paper, and Gabriele Bozzola for serving as editor. The views expressed in the article do not necessarily represent the views of the DOE or the U.S. Government. The example notebooks in Mantaray use and distribute data that were generated by the E.U. Copernicus Marine Service Information and the European Marine Observation and Data Network (EMODnet).
 
 # References
