### 2025-11-24 #8691 @biavillasboas
Hi @Sbozzolo, I believe we have addressed all items that you requested below. Please let me know if there are any additional changes you'd like to see.

- [Nazaré notebook](https://mines-oceanography.github.io/mantaray/nazare_example.html ) now renders the output from plots [see PR#244](https://github.com/mines-oceanography/mantaray/pull/244)
- Forking the repo is now recommended both for users and contributors, see [PR#245](https://github.com/mines-oceanography/mantaray/pull/245)
- DOIs have been included [see PR#243](https://github.com/mines-oceanography/mantaray/pull/243)
- Datasets are acknowledged in the paper acknowledgments following the language recommended by the data producers (note that all datasets have permissive licences) [see PR#246](https://github.com/mines-oceanography/mantaray/pull/246)

> Hi [@biavillasboas](https://github.com/biavillasboas), I have a couple of very small points before we can move to the post-review checklist:
>
> * Should https://mines-oceanography.github.io/mantaray/nazare_example.html show a plot?
> * For development, you suggest `git clone git@github.com:mines-oceanography/mantaray.git`. This requires a GitHub account (because you are using ssh instead of https) and won't allow external contributors to push to the repo (because they won't have permissions). Somewhere else you suggest forking the repo. Maybe you want to clarify a tiny bit your recommendations.
> * Could you fix the DOIs in the message above?
> * Could you make sure that all the datasets that you are going to distribute with Mantaray are acknowledged in the paper as well (If they need to be acknowledged)?
>
> Thank you very much!
