### 2025-11-24 #8691 @Sbozzolo
@biavillasboas, everything looks good on my end, so we can move to the final steps. Could you please have a look at the following tasks:

> - Double check authors and affiliations (including ORCIDs)
> - Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> - Archive the release on Zenodo/figshare/etc and post the DOI here.
> - Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> - Make sure that the license listed for the archive is the same as the software license.

Once you have cleared those, I'll complete my checklist and recommend your submission for acceptance.
