+++ .github/workflows/python-doc.yml
@@ -24,7 +24,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6
 
       - uses: prefix-dev/setup-pixi@v0.9.3
         with:
