### 2025-11-30 0fb04a4
doc: Adding DOI badge
