### 2025-11-30 a17940c
Try pandoc curly brackets convention
Trying to follow the pandoc convention and put "e.g.,"  in curly brackets
