### 2025-11-30 5c7d04b
doc: Adding missing Apache-2 license badge
