### 2025-11-30 6e8a8ea
DOI badge (#249)
* doc: Adding DOI badge

* doc: Adding missing Apache-2 license badge
