### 2025-11-30 1fcbd5e
Fix minor grammar typos in paper.md (#250)
