### 2025-11-30 128bc1d
cfg: Lock netcdf for now to 0.11.0
