+++ Cargo.toml
@@ -42,7 +42,7 @@ tracing = "0.1.41"
 tempfile = "3.20.0"
 
 [build-dependencies]
-cbindgen = { version = "0.28.0", optional = true }
+cbindgen = { version = "0.29.2", optional = true }
 pyo3-build-config = "0.25.0"
 
 [features]
