+++ Cargo.toml
@@ -31,7 +31,7 @@ ndarray = { version = "0.16.1", features = ["rayon"] }
 netcdf3 = "0.5.2"
 netcdf = {version = "=0.11.0", features = ["static", "ndarray"]}
 ode_solvers = "0.6.1"
-pyo3 = { version = "0.25.1", features = ["extension-module"] }
+pyo3 = { version = "0.27.2", features = ["extension-module"] }
 rayon = "1.10.0"
 serde_json = "1.0.141"
 serde = {version="1.0", features=["derive"]}
