### 2025-11-30 363d72e
cfg: Bump mantaray crate to 0.2.0
