+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mantaray"
-version = "0.1.0"
+version = "0.2.0"
 authors = [
   "Bryce Irving <bryceirving@mines.edu>",
   "Guilherme Castelao <guilherme@castelao.net>",
