### 2025-11-30 6bc1b39
cfg: Also bump pyo3-build-config to 0.27.2
