+++ Cargo.toml
@@ -43,7 +43,7 @@ tempfile = "3.20.0"
 
 [build-dependencies]
 cbindgen = { version = "0.29.2", optional = true }
-pyo3-build-config = "0.25.0"
+pyo3-build-config = "0.27.2"
 
 [features]
 capi = ["libc"]
