### 2025-11-30 250cce3
Update Rust side dependencies (#251)
* cfg: Updating dependencies

* cfg: Lock netcdf for now to 0.11.0

* cfg: Updating cbindgen

* cfg: Updating pyo3

* cfg: Bump mantaray crate to 0.2.0

* cfg: Also bump pyo3-build-config to 0.27.2
