+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mantaray"
-version = "0.1.0"
+version = "0.2.0"
 authors = [
   "Bryce Irving <bryceirving@mines.edu>",
   "Guilherme Castelao <guilherme@castelao.net>",
@@ -29,9 +29,9 @@ derive_builder = "0.20.0"
 libc = { version = "0.2.174", optional = true }
 ndarray = { version = "0.16.1", features = ["rayon"] }
 netcdf3 = "0.5.2"
-netcdf = {version = "0.11.0", features = ["static", "ndarray"]}
+netcdf = {version = "=0.11.0", features = ["static", "ndarray"]}
 ode_solvers = "0.6.1"
-pyo3 = { version = "0.25.1", features = ["extension-module"] }
+pyo3 = { version = "0.27.2", features = ["extension-module"] }
 rayon = "1.10.0"
 serde_json = "1.0.141"
 serde = {version="1.0", features=["derive"]}
@@ -42,8 +42,8 @@ tracing = "0.1.41"
 tempfile = "3.20.0"
 
 [build-dependencies]
-cbindgen = { version = "0.28.0", optional = true }
-pyo3-build-config = "0.25.0"
+cbindgen = { version = "0.29.2", optional = true }
+pyo3-build-config = "0.27.2"
 
 [features]
 capi = ["libc"]
