### 2025-11-30 b464350
fix: Locking on a working version of netcdf
