+++ Cargo.toml
@@ -30,6 +30,8 @@ libc = { version = "0.2.174", optional = true }
 ndarray = { version = "0.16.1", features = ["rayon"] }
 netcdf3 = "0.5.2"
 netcdf = {version = "=0.11.0", features = ["static", "ndarray"]}
+netcdf-src = {version = "=0.4.1" }
+netcdf-sys = {version = "=0.8.1" }
 ode_solvers = "0.6.1"
 pyo3 = { version = "0.27.2", features = ["extension-module"] }
 rayon = "1.10.0"
