### 2025-12-01 7f5241b
cfg: Update Rust used with pixi
