### 2025-12-01 6843cd6
cfg: Bump uv version
