### 2025-12-01 20fb449
cfg: Bump ruff version
