### 2025-12-01 63270a2
cfg: Bump Mantaray (Python) to 0.2.1
