### 2025-12-01 bb8ec56
Updating Python side dependencies (#253)
* cfg: Update Rust used with pixi

* cfg: Updating Maturin

* cfg: Bump uv version

* cfg: Bump ruff version

* cfg: Updating dependencies in pixi

* cfg: Bump Mantaray (Python) to 0.2.1
