### 2025-12-01 61c366d
cfg: Creating a CITATION.cff
