+++ CITATION.cff
@@ -1,0 +1,52 @@
+cff-version: 1.2.0
+title: >-
+  Mantaray: A Rust Package for Ray Tracing Ocean Surface
+  Gravity Waves
+message: >-
+  If you use this software, please cite it using the
+  metadata from this file.
+type: software
+authors:
+  - given-names: Bryce
+    family-names: Irving
+    affiliation: Colorado School of Mines
+    orcid: 'https://orcid.org/0009-0004-2309-9522'
+  - given-names: Guilherme
+    family-names: Castelao
+    affiliation: National Renewable Energy Laboratory
+    orcid: 'https://orcid.org/0000-0002-6765-0708'
+  - given-names: Colin
+    family-names: Beyers
+    affiliation: Colorado School of Mines
+    orcid: 'https://orcid.org/0009-0004-8312-6158'
+  - given-names: James
+    family-names: Clemson
+    affiliation: Colorado School of Mines
+    orcid: 'https://orcid.org/0009-0000-4329-6575'
+  - given-names: Jackson
+    family-names: Krieger
+    affiliation: Colorado School of Mines
+    orcid: 'https://orcid.org/0009-0006-3693-8887'
+  - given-names: Gwendal
+    family-names: Marechal
+    affiliation: Colorado School of Mines
+    orcid: 'https://orcid.org/0000-0003-0378-5694'
+  - given-names: Nicholas
+    family-names: Pizzo
+    affiliation: >-
+      Graduate School of Oceanography, University of Rhode
+      Island
+    orcid: 'https://orcid.org/0000-0001-9570-4200'
+  - given-names: Bia
+    family-names: Villas Bôas
+    affiliation: Colorado School of Mines
+    orcid: 'https://orcid.org/0000-0001-6767-6556'
+repository-code: 'https://github.com/mines-oceanography/mantaray'
+url: 'https://github.com/mines-oceanography/mantaray'
+keywords:
+  - ray tracing
+  - waves
+  - oceanography
+license: Apache-2.0
+version: 0.2.1
+date-released: '2025-11-30'
