### 2025-12-02 224fc5c
Releasing 0.3.0 as JOSS publication
