+++ CITATION.cff
@@ -48,5 +48,5 @@ keywords:
   - waves
   - oceanography
 license: Apache-2.0
-version: 0.2.1
-date-released: '2025-11-30'
+version: 0.3.0
+date-released: '2025-12-01'
