### 2025-12-02 2d30ff5
Releasing 0.3.0 as JOSS publication (#257)
