### 2025-12-02 #8691 @biavillasboas
Hi, @Sbozzolo

✅ Double check authors and affiliations (including ORCIDs)
✅ Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper: [latest release is v0.3.0](https://github.com/mines-oceanography/mantaray/tree/v0.3.0)
✅ Archive the release on Zenodo/figshare/etc and post the DOI here: Archived on Zenodo with [DOI: 10.5281/zenodo.17133200](https://doi.org/10.5281/zenodo.17133200)
✅  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
✅ Make sure that the license listed for the archive is the same as the software license.

Please let me know if you need anything else from my end.
