### 2025-12-05 #8691 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "delaware" but please check the rest.
