### 2025-12-05 d0883bc
Fix capitalization in article titles in paper.bib
