### 2025-12-05 65f2a3d
Fix capitalization in book title entry
