+++ joss/paper.bib
@@ -103,7 +103,7 @@
 }
 
 @book{phillips1966,
-  title = {The {Dynamics} of the {Upper} {Ocean}},
+  title = {The dynamics of the upper ocean},
   language = {en},
   urldate = {2024-11-14},
   publisher = {Cambridge University Press},
