### 2025-12-06 78b63fc
Fix capitalization on citations (#258)
* Try pandoc curly brackets convention

Trying to follow the pandoc convention and put "e.g.,"  in curly brackets

* Try in-text citation instead

Try to write down citations keys in-text

* Test removing comma after e.g.

* Change order of prefix to test alphabetical order

* Back to original version

* Fix capitalization in article titles in paper.bib

* Fix capitalization in book title entry
