+++ joss/paper.bib
@@ -33,7 +33,7 @@
 }
 
 @article{oreilly2016california,
-  title={The California coastal wave monitoring and prediction system},
+  title={The {C}alifornia coastal wave monitoring and prediction system},
   author={O'Reilly, William C and Olfe, Corey B and Thomas, Julianna and Seymour, RJ and Guza, RT},
   journal={Coastal Engineering},
   volume={116},
@@ -91,7 +91,7 @@
 }
 
 @article{kukulka2017surface,
-  title={Surface wave dynamics in Delaware Bay and its adjacent coastal shelf},
+  title={Surface wave dynamics in {D}elaware {B}ay and its adjacent coastal shelf},
   author={Kukulka, Tobias and Jenkins III, Robert L and Kirby, James T and Shi, Fengyan and Scarborough, Robert W},
   journal={Journal of Geophysical Research: Oceans},
   volume={122},
@@ -103,7 +103,7 @@
 }
 
 @book{phillips1966,
-  title = {The {Dynamics} of the {Upper} {Ocean}},
+  title = {The dynamics of the upper ocean},
   language = {en},
   urldate = {2024-11-14},
   publisher = {Cambridge University Press},
